chore: move src to packages
