+++ packages/functions/src/Reduce.ts
