+++ packages/loops/src/helpers/interruption.ts
