+++ packages/switch/src/Switch.ts
