--- src/index.ts
@@ -1,29 +1,0 @@
-export { If } from './Conditions/If';
-export { Else } from './Conditions/Else';
-export { Then } from './Conditions/Then';
-export { When } from './Conditions/When';
-export { Unless } from './Conditions/Unless';
-
-export { Map } from './Functions/Map';
-export { Filter } from './Functions/Filter';
-export { Reduce } from './Functions/Reduce';
-
-export { For } from './Loops/For';
-export { ForOf } from './Loops/ForOf';
-export { ForIn } from './Loops/ForIn';
-export { ForEach } from './Loops/ForEach';
-export { While } from './Loops/While';
-export { DoWhile } from './Loops/DoWhile';
-export { Break } from './Loops/Break';
-export { Continue } from './Loops/Continue';
-
-export { RPromise } from './Promises/Promise';
-export { Resolve } from './Promises/Resolve';
-export { Reject } from './Promises/Reject';
-
-export { Switch } from './Switch/Switch';
-export { Case } from './Switch/Case';
-export { Default } from './Switch/Default';
-
-export { SetTimeout } from './Timeouts/SetTimeout';
-export { SetInterval } from './Timeouts/SetInterval';
